requirements.txt
python-dotenv
groq
langchain
langchain-community
faiss-cpu
pymupdf
sentence-transformers
